python,runfiles: move to different package
Move the Python runfiles library from
`@bazel_tools//tools/runfiles:py-runfiles` to
`@bazel_tools//tools/python/runfiles:runfiles`

Also rename the testdata runfiles.py to foo.py.
This file was not a mock runfiles library, just a
client file using the runfiles library that was
also called runfiles.py

Fixes https://github.com/bazelbuild/bazel/issues/4878


Closes #4981.

PiperOrigin-RevId: 192256481
