+++ src/test/py/bazel/runfiles_test.py
@@ -74,7 +74,7 @@ class RunfilesTest(test_base.TestBase):
     for s, t in [
         ("WORKSPACE.mock", "WORKSPACE"),
         ("foo/BUILD.mock", "foo/BUILD"),
-        ("foo/runfiles.py", "foo/runfiles.py"),
+        ("foo/foo.py", "foo/foo.py"),
         ("foo/datadep/hello.txt", "foo/datadep/hello.txt"),
         ("bar/BUILD.mock", "bar/BUILD"),
         ("bar/bar.py", "bar/bar.py"),
@@ -131,19 +131,15 @@ class RunfilesTest(test_base.TestBase):
 
       i += 2
 
-  # TODO(laszlocsomor): re-enable after
-  # https://github.com/bazelbuild/bazel/issues/4878 is fixed.
-  # def testPythonRunfilesLibraryInBazelToolsRepo(self):
-  #  self._AssertPythonRunfilesLibraryInBazelToolsRepo("py", "Python")
+  def testPythonRunfilesLibraryInBazelToolsRepo(self):
+    self._AssertPythonRunfilesLibraryInBazelToolsRepo("py", "Python")
 
   def testRunfilesLibrariesFindRunfilesWithoutEnvvars(self):
     for s, t in [
         ("WORKSPACE.mock", "WORKSPACE"),
         ("bar/BUILD.mock", "bar/BUILD"),
-        # TODO(laszlocsomor): uncomment Python files after
-        # https://github.com/bazelbuild/bazel/issues/4878 is fixed.
-        # ("bar/bar.py", "bar/bar.py"),
-        # ("bar/bar-py-data.txt", "bar/bar-py-data.txt"),
+        ("bar/bar.py", "bar/bar.py"),
+        ("bar/bar-py-data.txt", "bar/bar-py-data.txt"),
         ("bar/Bar.java", "bar/Bar.java"),
         ("bar/bar-java-data.txt", "bar/bar-java-data.txt"),
     ]:
@@ -155,12 +151,11 @@ class RunfilesTest(test_base.TestBase):
     self.AssertExitCode(exit_code, 0, stderr)
     bazel_bin = stdout[0]
 
-    exit_code, _, stderr = self.RunBazel(["build", "//bar:bar-java"])
+    exit_code, _, stderr = self.RunBazel(
+        ["build", "//bar:bar-py", "//bar:bar-java"])
     self.AssertExitCode(exit_code, 0, stderr)
 
-    # TODO(laszlocsomor): add Python after
-    # https://github.com/bazelbuild/bazel/issues/4878 is fixed.
-    for lang in [("java", "Java", "Bar.java")]:
+    for lang in [("py", "Python", "bar.py"), ("java", "Java", "Bar.java")]:
       if test_base.TestBase.IsWindows():
         bin_path = os.path.join(bazel_bin, "bar/bar-%s.exe" % lang[0])
       else:
