+++ src/test/py/bazel/testdata/runfiles_test/bar/BUILD.mock
@@ -5,7 +5,7 @@ py_binary(
     srcs = ["bar.py"],
     data = ["bar-py-data.txt"],
     main = "bar.py",
-    deps = ["@bazel_tools//tools/runfiles:py-runfiles"],
+    deps = ["@bazel_tools//tools/python/runfiles"],
 )
 
 java_binary(
