+++ src/tools/runfiles/BUILD
@@ -22,22 +22,9 @@ filegroup(
         "BUILD.tools",
         "runfiles.cc",
         "runfiles.h",
-        "runfiles.py",
         "//src/tools/runfiles/java/com/google/devtools/build/runfiles:embedded_tools",
     ],
     visibility = ["//src:__pkg__"],
-)
-
-py_library(
-    name = "py-runfiles",
-    srcs = ["runfiles.py"],
-)
-
-py_test(
-    name = "py-runfiles-test",
-    srcs = ["runfiles_test.py"],
-    main = "runfiles_test.py",
-    deps = [":py-runfiles"],
 )
 
 cc_library(
